+++ include/pulse/Pulse.Lib.C.fsti
@@ -1,4 +1,5 @@
 module Pulse.Lib.C
+include FStar.Mul
 include Pulse.Lib.C.Inhabited
 include Pulse.Lib.C.Int32
 include Pulse.Lib.C.Ref
